api token

+++ wannapop/api/orders.py
@@ -1,4 +1,5 @@
 from flask import request, jsonify
+from .helper_auth import basic_auth, token_auth
 from . import api_bp
 from ..models import Order, ConfirmedOrder
 from .. import db_manager as db
@@ -33,22 +34,47 @@ def anular_oferta(id):
         return jsonify({'error': 'Bad Request', 'message': 'No se pudo anular la oferta', 'success': False}), 400
 
 @api_bp.route('/orders/<int:id>/confirmed', methods=['POST'])
+@token_auth.login_required
 def aceptar_oferta(id):
+    current_user = token_auth.current_user()
+    current_user_id = current_user.id
+
+    # Obtener el producto asociado con la orden
     order = Order.query.get_or_404(id)
-    
-    # Creamos una instancia de ConfirmedOrder utilizando el método create del mixin
-    try:
-        confirmed_order = ConfirmedOrder.create(order=order)
-        return jsonify({'message': 'Oferta aceptada exitosamente', 'success': True}), 200
-    except Exception as e:
-        return jsonify({'error': 'Internal Server Error', 'message': str(e), 'success': False}), 500
+    product = order.product
+
+    # Verificar si el usuario actual es el vendedor del producto
+    if current_user_id == product.seller_id:
+        try:
+            # Creamos una instancia de ConfirmedOrder utilizando el método create del mixin
+            confirmed_order = ConfirmedOrder.create(order=order)
+            return jsonify({'message': 'Oferta aceptada exitosamente', 'success': True}), 200
+        except Exception as e:
+            return jsonify({'error': 'Internal Server Error', 'message': str(e), 'success': False}), 500
+    else:
+        return jsonify({'error': 'Forbidden', 'message': 'No tiene permiso para aceptar esta oferta', 'success': False}), 403
+
 
 @api_bp.route('/orders/<int:id>/confirmed', methods=['DELETE'])
+@token_auth.login_required
 def anular_oferta_aceptada(id):
-    confirmed_order = ConfirmedOrder.query.filter_by(order_id=id).first_or_404()
-    
-    try:
-        confirmed_order.delete()
-        return jsonify({'message': 'Oferta aceptada anulada exitosamente', 'success': True}), 200
-    except Exception as e:
-        return jsonify({'error': 'Internal Server Error', 'message': str(e), 'success': False}), 500
+    current_user = token_auth.current_user()
+    current_user_id = current_user.id
+
+    # Obtener la orden correspondiente
+    order = Order.query.get_or_404(id)
+
+    # Obtener el producto asociado con la orden
+    product = order.product
+
+    # Verificar si el usuario actual es el vendedor del producto
+    if current_user_id == product.seller_id:
+        confirmed_order = ConfirmedOrder.query.filter_by(order_id=id).first_or_404()
+        
+        try:
+            confirmed_order.delete()
+            return jsonify({'message': 'Oferta aceptada anulada exitosamente', 'success': True}), 200
+        except Exception as e:
+            return jsonify({'error': 'Internal Server Error', 'message': str(e), 'success': False}), 500
+    else:
+        return jsonify({'error': 'Forbidden', 'message': 'No tiene permiso para anular esta oferta', 'success': False}), 403
